split views into package
