+++ api/views/__init__.py
@@ -1,0 +1,2 @@
+from .account import *  # noqa
+from .message import *  # noqa
